+++ lib/puppet_x/cf_web/docker.rb
@@ -1,0 +1,7 @@
+#
+# Copyright 2019 (c) Andrey Galkin
+#
+
+
+module PuppetX::CfWeb::Docker
+end