+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - NEW: backend auto-discovery support
 - NEW: "unlimited" placeholder for site limits
 - NEW: cfbackup integration for persistent directories
+- NEW: experimental Docker support
 
 # 1.2.0 (2018-12-09)
 - CHANGED: updated for Ubuntu 18.04 Bionic support